refactor: mover constantes de domínio para src/lib/constants.js
Resolve #52. Elimina acoplamento página→página.

Constantes movidas:
- SERVICES_CONFIG (era export de NewOnboarding.jsx)
- BUSINESS_LABELS, MATURITY_LABELS, CONTRACT_MODEL_LABELS, CONTRACT_PAYMENT_LABELS
- EDIT_BUSINESS_TYPES, EDIT_MATURITY_OPTIONS
- SEGMENTOS (unificado com EDIT_SEGMENTOS — eram listas idênticas)

Arquivos atualizados para importar de src/lib/constants:
NewOnboarding, ClientProfile, OnboardingEditForm, OnboardingContent.
